pyproject.toml
[project]
name = "wochenendaufgabe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "django>=6.0",
    "django-service-objects>=0.7.1",
    "fastmcp>=2.14.0",
    "mysql-connector-python>=9.5.0",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
